wupper_module: Add DMA descriptors enum
Add a enum that defines constants for the indexes of the DMA
descriptors. This way, the user can pass an enum member to the
dma_to_host() call instead of a literal. This approach is advised in the
method documentation. Also fixes the dma_index parameter of that method
to receive an int instead of an uint8_t.

+++ wupper_module/wupper.cpp
@@ -43,7 +43,7 @@ Wupper::Wupper(const std::string device_name, int iommu_group_id)
 }
 
 void Wupper::dma_to_host(
-    size_t transfer_size, uint8_t dma_index, DataBuffer &buffer)
+    size_t transfer_size, int dma_index, DataBuffer &buffer)
 {
     volatile wuppercard_bar0_regs_t *bar0_regs
         = (volatile wuppercard_bar0_regs_t *)bar0.data;
